[Dev] suspicious: implement the suspicious radius for the first layer in proximity computation

+++ api/models/request_next_data.go
@@ -12,7 +12,7 @@ var (
 )
 
 type RequestNextSuspicionData struct {
-	Radius int `json:"radius"` // 0 = disable
+	Radius uint8 `json:"radius"` // 0 = disable
 }
 
 type RequestNextDepthData struct {
